+++ inst/shiny/server.R
@@ -192,9 +192,11 @@ server <- function(input, output) {
   data_plot <- reactive({
     req(data_ac())
     if(input$data_omit) {
-      dplyr::bind_rows(data_ac()[1,],
-                       dplyr::filter(data_ac()[-1,],
-                                     abs(as.numeric(.data$charge_balance)) < 10))
+
+      dplyr::bind_rows(
+        data_ac()[1,],
+        dplyr::filter(data_ac()[-1,],
+                      abs(as.numeric(.data[[input$data_charge_balance]])) < 10))
     } else data_ac()
   })
 
@@ -232,11 +234,15 @@ server <- function(input, output) {
                                          "Ca_meq", "Mg_meq", "Na_meq", "Cl_meq",
                                          "HCO3_meq", "SO4_meq",
                                          "cation_sum", "anion_sum",
-                                         "charge_balance")))
+                                         "charge_balance",
+                                         "cation_sum2", "anion_sum2",
+                                         "charge_balance2")))
 
     }
     d <- dplyr::select(d, "StationID", "SampleID", "Sample_Date", "cation_sum",
-                       "anion_sum", "charge_balance", dplyr::everything()) %>%
+                       "anion_sum", "charge_balance",
+                       "cation_sum2", "anion_sum2", "charge_balance2",
+                       dplyr::everything()) %>%
       dplyr::arrange("StationID", "SampleID", "Sample_Date")
 
     col_names <- paste(colnames(d), units[colnames(d)], sep = "\n")
@@ -246,8 +252,10 @@ server <- function(input, output) {
                   rownames = FALSE) %>%
       DT::formatRound(columns = c("Ca_meq", "Mg_meq", "Na_meq", "Cl_meq",
                                   "SO4_meq", "HCO3_meq", "cation_sum",
-                                  "anion_sum", "charge_balance")) %>%
-      DT::formatStyle("charge_balance",
+                                  "anion_sum", "charge_balance",
+                                  "cation_sum2", "anion_sum2",
+                                  "charge_balance2")) %>%
+      DT::formatStyle(input$data_charge_balance,
                       backgroundColor = DT::styleInterval(
                         cuts = c(-10, 10),
                         values = c("#f8d7da", "#d4edda", "#f8d7da")))